+++ packages/charting/configure/src/configure.jsx
@@ -59,7 +59,8 @@ export class Configure extends React.Component {
 
   constructor(props) {
     super(props);
-    const { range, graph } = props.model || {};
+    const { range = {}, graph } = props.model || {};
+
 
     const gridValues = {
       range: getGridValues(range, graph.height, true)
@@ -127,7 +128,7 @@ export class Configure extends React.Component {
       labelsPlaceholders = {},
       titlePlaceholder = {},
       chartDimensions = {},
-      withRubric,
+      withRubric = {},
     } = configuration || {};
     const {
       teacherInstructionsEnabled,
@@ -168,7 +169,7 @@ export class Configure extends React.Component {
                     'all or nothing',
                     'partial scoring',
                   ]),
-                rubricEnabled: withRubric.settings && toggle(withRubric.label)
+                rubricEnabled: withRubric?.settings && toggle(withRubric?.label)
               },
             }}
           />