+++ packages/extended-text-entry/configure/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@6.3.0...@pie-element/extended-text-entry-configure@6.3.1) (2022-11-11)
+
+
+### Bug Fixes
+
+* **extended-text-entry:** add sanity checks for withRubric ([01ee1db](https://github.com/pie-framework/pie-elements/commit/01ee1db32ec5f6d6fc2602ee5d25ad5204ede456))
+
+
+
+
+
+# [6.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@6.2.2...@pie-element/extended-text-entry-configure@6.3.0) (2022-11-01)
+
+
+### Features
+
+* add a different property in the config, forceEnabled, to force a value ([300c664](https://github.com/pie-framework/pie-elements/commit/300c664618f46177137deef3bced2d096cdb8126))
+
+
+
+
+
+## [6.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@6.2.1...@pie-element/extended-text-entry-configure@6.2.2) (2022-11-01)
+
+
+### Bug Fixes
+
+* trigger updating model depending on configuration.withRubric ([6b42963](https://github.com/pie-framework/pie-elements/commit/6b429639dade1f72e77e972530ab784c5417384c))
+* trigger updating model depending on configuration.withRubric (added additional check) ([02ae9c0](https://github.com/pie-framework/pie-elements/commit/02ae9c0664aa140d6624643b6efbdef33cea2acb))
+
+
+
+
+
+## [6.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@6.2.0...@pie-element/extended-text-entry-configure@6.2.1) (2022-10-31)
+
+**Note:** Version bump only for package @pie-element/extended-text-entry-configure
+
+
+
+
+
 # [6.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/extended-text-entry-configure@6.1.9...@pie-element/extended-text-entry-configure@6.2.0) (2022-10-17)