+++ packages/image-cloze-association/configure/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@4.2.1...@pie-element/image-cloze-association-configure@4.2.2) (2022-11-11)
+
+
+### Bug Fixes
+
+* **image-close-association:** add sanity checks for withRubric ([1d3c577](https://github.com/pie-framework/pie-elements/commit/1d3c577d04b0c9b13589c9fce901b673be27ad31))
+
+
+
+
+
+## [4.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@4.2.0...@pie-element/image-cloze-association-configure@4.2.1) (2022-10-31)
+
+**Note:** Version bump only for package @pie-element/image-cloze-association-configure
+
+
+
+
+
 # [4.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/image-cloze-association-configure@4.1.9...@pie-element/image-cloze-association-configure@4.2.0) (2022-10-17)