+++ packages/inline-dropdown/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown@4.4.0...@pie-element/inline-dropdown@4.4.1) (2022-11-11)
+
+
+### Bug Fixes
+
+* **inline-dropdown:** add sanity checks for withRubric ([25e5403](https://github.com/pie-framework/pie-elements/commit/25e540323dd33c43f67e2f8ade736a4fc98c96f2))
+
+
+
+
+
+# [4.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown@4.3.1...@pie-element/inline-dropdown@4.4.0) (2022-11-01)
+
+
+### Features
+
+* add a different property in the config, forceEnabled, to force a value ([300c664](https://github.com/pie-framework/pie-elements/commit/300c664618f46177137deef3bced2d096cdb8126))
+
+
+
+
+
+## [4.3.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown@4.3.0...@pie-element/inline-dropdown@4.3.1) (2022-10-31)
+
+
+### Bug Fixes
+
+* **inline-dropdown:** fix count in choice rationale PD-2056 ([37bebe1](https://github.com/pie-framework/pie-elements/commit/37bebe168eb4ee4f1f7aad4330e548344d46f308))
+
+
+
+
+
 # [4.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/inline-dropdown@4.2.5...@pie-element/inline-dropdown@4.3.0) (2022-10-17)