+++ packages/likert/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/likert@0.3.1...@pie-element/likert@0.4.0) (2022-11-01)
+
+
+### Features
+
+* add a different property in the config, forceEnabled, to force a value ([300c664](https://github.com/pie-framework/pie-elements/commit/300c664618f46177137deef3bced2d096cdb8126))
+
+
+
+
+
+## [0.3.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/likert@0.3.0...@pie-element/likert@0.3.1) (2022-10-31)
+
+**Note:** Version bump only for package @pie-element/likert
+
+
+
+
+
 # [0.3.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/likert@0.2.10...@pie-element/likert@0.3.0) (2022-10-17)